requirements/develop.txt
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
channels==1.0.3
click==6.7
constantly==15.1.0
daphne==1.0.2
decorator==4.0.6
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.30.0
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-avatar==3.1.0
django-badges==0.2.0
django-el-pagination==3.0.1
django-emoji==2.2.0
django-extensions==1.7.6
django-messages==0.5.3
django-model-utils==2.6.1
django-notifications-hq==1.2
django-sortedm2m==1.3.3
django-taggit==0.22.0
djangorestframework==3.5.3
ExifRead==2.1.2
gunicorn==19.6.0
incremental==16.10.1
ipdb==0.10.2
ipython==5.2.1
ipython-genutils==0.1.0
jsonfield==1.0.3
msgpack-python==0.4.8
oauthlib==2.0.1
olefile==0.44
packaging==16.8
path.py==10.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.10
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
sockjs-tornado==1.0.3
SwampDragon==0.4.2.2
tornado==4.4.2
tornado-redis==2.4.18
traitlets==4.3.1
Twisted==16.6.0
txaio==2.6.0
wcwidth==0.1.7
Werkzeug==0.11.15
zope.interface==4.3.3
asgi_rabbitmq
django-celery-results
